Replace GREP_OPTIONS with a script
While GREP_OPTIONS was dangerous, it could be safely used within
confined contexts. However, it's no longer supported:
    https://lists.gnu.org/archive/html/grep-commit/2014-09/msg00007.html

Unfortunately, the only way to pass options is directly on the
command-line. The quickfix window title displays the grepprg used, and
we use so many options it would be hard to see what you searched for.
Use a script to avoid the options obscuring your query in the quickfix
window title.