Always invoke SmartGrepToggle so we set grepprg
Now that we're using a script for smart grep instead of options and
we're switching the value of grepprg to determine smartness (instead of
variables), we need to always run SmartGrepToggle. Use the value of
searchsavvy_smartgrep_auto_enable to determine the value passed to
SmartGrepToggle so we still support disabled by default -- but disabled
will still use grep and not findstr on Windows.